feat: add fixtures module to hold sample data for tests and previews
