+++ core/data/src/test/kotlin/com/divinelink/core/data/details/person/ProdPersonRepositoryTest.kt
@@ -7,6 +7,8 @@ import com.divinelink.core.data.person.details.mapper.map
 import com.divinelink.core.data.person.repository.ProdPersonRepository
 import com.divinelink.core.database.person.PersonDao
 import com.divinelink.core.database.person.ProdPersonDao
+import com.divinelink.core.fixtures.core.commons.ClockFactory
+import com.divinelink.core.fixtures.model.person.credit.PersonCombinedCreditsFactory
 import com.divinelink.core.network.changes.model.api.ChangesResponseApi
 import com.divinelink.core.network.client.localJson
 import com.divinelink.core.network.details.person.model.PersonCreditsApi
@@ -14,12 +16,10 @@ import com.divinelink.core.network.media.model.changes.ChangesParameters
 import com.divinelink.core.testing.MainDispatcherRule
 import com.divinelink.core.testing.database.TestDatabaseFactory
 import com.divinelink.core.testing.factories.api.details.person.PersonDetailsApiFactory
-import com.divinelink.core.testing.factories.core.commons.ClockFactory
 import com.divinelink.core.testing.factories.entity.person.PersonEntityFactory
 import com.divinelink.core.testing.factories.entity.person.credits.PersonCastCreditEntityFactory
 import com.divinelink.core.testing.factories.entity.person.credits.PersonCrewCreditEntityFactory
 import com.divinelink.core.testing.factories.model.change.ChangeSample
-import com.divinelink.core.testing.factories.model.person.credit.PersonCombinedCreditsFactory
 import com.divinelink.core.testing.service.TestPersonService
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
