+++ core/domain/src/main/kotlin/com/divinelink/core/domain/FetchWatchlistUseCase.kt
@@ -1,5 +1,6 @@
 package com.divinelink.core.domain
 
+import com.divinelink.core.commons.domain.DispatcherProvider
 import com.divinelink.core.commons.domain.FlowUseCase
 import com.divinelink.core.data.account.AccountRepository
 import com.divinelink.core.data.session.model.SessionException
@@ -7,7 +8,6 @@ import com.divinelink.core.datastore.SessionStorage
 import com.divinelink.core.model.media.MediaType
 import com.divinelink.core.model.watchlist.WatchlistParameters
 import com.divinelink.core.model.watchlist.WatchlistResponse
-import com.divinelink.core.commons.domain.DispatcherProvider
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
