+++ core/domain/src/main/kotlin/com/divinelink/core/domain/session/CreateSessionUseCase.kt
@@ -1,11 +1,11 @@
 package com.divinelink.core.domain.session
 
+import com.divinelink.core.commons.domain.DispatcherProvider
 import com.divinelink.core.commons.domain.UseCase
 import com.divinelink.core.commons.domain.data
 import com.divinelink.core.data.session.repository.SessionRepository
 import com.divinelink.core.datastore.SessionStorage
 import com.divinelink.core.network.session.model.CreateSessionRequestApi
-import com.divinelink.core.commons.domain.DispatcherProvider
 
 class CreateSessionUseCase(
   private val repository: SessionRepository,
