+++ core/domain/src/main/kotlin/com/divinelink/core/domain/session/LogoutUseCase.kt
@@ -1,9 +1,9 @@
 package com.divinelink.core.domain.session
 
+import com.divinelink.core.commons.domain.DispatcherProvider
 import com.divinelink.core.commons.domain.UseCase
 import com.divinelink.core.data.session.repository.SessionRepository
 import com.divinelink.core.datastore.SessionStorage
-import com.divinelink.core.commons.domain.DispatcherProvider
 
 class LogoutUseCase(
   private val repository: SessionRepository,
