+++ core/fixtures/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+  alias(libs.plugins.divinelink.android.library)
+  alias(libs.plugins.divinelink.android.library.compose)
+}
+
+dependencies {
+  api(projects.core.model)
+
+  implementation(libs.kotlinx.datetime)
+}
