+++ core/testing/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 dependencies {
   api(projects.core.data)
   api(projects.core.model)
+  api(projects.core.fixtures)
 
   api(libs.compose.ui.test.junit4)
   api(libs.kotlin.test.junit)
