+++ feature/details/src/main/kotlin/com/divinelink/feature/details/person/ui/PersonContent.kt
@@ -56,7 +56,7 @@ fun PersonContent(
 
 @Previews
 @Composable
-private fun PersonContentPreview(
+fun PersonContentPreview(
   @PreviewParameter(PersonUiStatePreviewParameterProvider::class)
   uiState: PersonUiState,
 ) {
