+++ feature/details/src/main/kotlin/com/divinelink/feature/details/person/ui/provider/PersonUiStatePreviewParameterProvider.kt
@@ -1,51 +1,30 @@
 package com.divinelink.feature.details.person.ui.provider
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import androidx.compose.ui.tooling.preview.datasource.LoremIpsum
-import com.divinelink.core.model.credits.PersonRole
-import com.divinelink.core.model.details.Person
-import com.divinelink.core.model.details.person.PersonDetails
-import com.divinelink.core.model.person.Gender
+import com.divinelink.core.fixtures.details.person.PersonDetailsFactory
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory
 import com.divinelink.feature.details.person.ui.PersonDetailsUiState
 import com.divinelink.feature.details.person.ui.PersonUiState
 
 class PersonUiStatePreviewParameterProvider : PreviewParameterProvider<PersonUiState> {
   override val values: Sequence<PersonUiState> = sequenceOf(
     PersonUiState(
-      personDetails = PersonDetailsUiState.Data.Visible(PersonUiStateData.steveCarell),
+      personDetails = PersonDetailsUiState.Data.Visible(PersonDetailsFactory.steveCarell()),
     ),
     // Sorry for making you dead Steve Carell, it's only for testing purposes!
     PersonUiState(
       personDetails = PersonDetailsUiState.Data.Visible(
-        PersonUiStateData.steveCarell.copy(deathday = "2022-05-16"),
+        PersonDetailsFactory.steveCarell().copy(deathday = "2022-05-16"),
       ),
     ),
     PersonUiState(
       personDetails = PersonDetailsUiState.Data.Visible(
-        PersonUiStateData.steveCarell.copy(biography = null),
+        PersonDetailsFactory.steveCarell().copy(biography = null),
       ),
     ),
+    PersonUiState(
+      personDetails = PersonDetailsUiState.Data.Visible(PersonDetailsFactory.steveCarell()),
+      credits = PersonCastCreditFactory.all(),
+    ),
   )
-
-  private object PersonUiStateData {
-    val steveCarell = PersonDetails(
-      person = Person(
-        id = 4495,
-        name = "Steve carell",
-        profilePath = "/dzJtsLspH5Bf8Tvw7OQC47ETNfJ.jpg",
-        gender = Gender.MALE,
-        role = listOf(PersonRole.Unknown),
-        knownForDepartment = "Acting",
-      ),
-      biography = LoremIpsum(50).values.joinToString(),
-      birthday = "1962-08-16",
-      deathday = null,
-      placeOfBirth = "Concord, Massachusetts, USA",
-      homepage = null,
-      alsoKnownAs = emptyList(),
-      imdbId = "nm0136797",
-      popularity = 77.108,
-      insertedAt = "1642185600",
-    )
-  }
 }
