+++ feature/details/src/test/kotlin/com/divinelink/feature/details/person/ui/PersonScreenTest.kt
@@ -16,13 +16,13 @@ import com.divinelink.core.model.media.MediaType
 import com.divinelink.core.navigation.arguments.DetailsNavArguments
 import com.divinelink.core.navigation.arguments.PersonNavArguments
 import com.divinelink.core.testing.ComposeTest
-import com.divinelink.core.testing.factories.details.person.PersonDetailsFactory
-import com.divinelink.core.testing.factories.model.person.credit.PersonCastCreditFactory
-import com.divinelink.core.testing.factories.model.person.credit.PersonCastCreditFactory.bruceAlmighty
-import com.divinelink.core.testing.factories.model.person.credit.PersonCastCreditFactory.despicableMe
-import com.divinelink.core.testing.factories.model.person.credit.PersonCastCreditFactory.littleMissSunshine
-import com.divinelink.core.testing.factories.model.person.credit.PersonCastCreditFactory.theOffice
-import com.divinelink.core.testing.factories.model.person.credit.PersonCombinedCreditsFactory
+import com.divinelink.core.fixtures.details.person.PersonDetailsFactory
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory.bruceAlmighty
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory.despicableMe
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory.littleMissSunshine
+import com.divinelink.core.fixtures.model.person.credit.PersonCastCreditFactory.theOffice
+import com.divinelink.core.fixtures.model.person.credit.PersonCombinedCreditsFactory
 import com.divinelink.core.testing.navigator.FakeDestinationsNavigator
 import com.divinelink.core.testing.setContentWithTheme
 import com.divinelink.core.testing.usecase.TestFetchChangesUseCase
